refactor(tools): dedupe PLY parser into ply_parse_core (#393)
Prior to this change, the engine-side `gseurat::ply::load`
(`src/engine/gaussian_cloud_ply.cpp`) and the standalone `ply_importer`
(`tools/ply_importer/ply_parse.cpp`) carried ~270 LOC of near-identical
PLY-format code: header walk, property aliasing, SH→RGB transform,
sigmoid-on-opacity, scale exp(), rotation normalize, and the bone_index
uchar fast path. PR #392 chose this duplication over coupling the
standalone tool to `gseurat_core`. #393 tracked the drift risk.

This change extracts the common core into a new `ply_parse_core` static
library at the repository root. Both consumers become thin wrappers
(~30 LOC each) that convert `RawSplat` rows into their target struct.

Architecture preserved:
- `ply_parse_core` depends ONLY on glm + stdlib. No `gseurat_core` dep.
- Runtime engine library still does not compile `gaussian_cloud_ply.cpp`
  and still hits the GSVX fast path via `load_with_gsvx_first`.
- `tools/ply_importer/` still does not link `gseurat_core`; only adds
  the new `ply_parse_core` lib.
- Engine-side `Gaussian` (per-field, 68 B) vs tool-side `GpuGaussian`
  (4×vec4, 64 B, std430) layouts stay consumer-specific. The shared
  intermediate `RawSplat` is parser-internal.

Wiring:
- New `ply_parse_core/` directory with CMakeLists, header, cpp.
- Top-level `CMakeLists.txt`: `add_subdirectory(ply_parse_core)` early,
  `add_gseurat_test` helper appends `ply_parse_core` to every test's
  link line so the 10+ tests that compile `gaussian_cloud_ply.cpp`
  link correctly (no-op for tests that don't reference any symbols).
- `tools/ply_importer/CMakeLists.txt`: add `ply_parse_core` to
  `target_link_libraries`.
- `ply2heightmap`: same.

Single behavior delta caught during review:
- Empty PLY (`vertex 0`) handling: the legacy engine parser silently
  produced an empty cloud; the legacy tool parser threw. The new
  shared parser preserves the engine behavior (returns empty
  `RawPlyCloud`) since `test_gaussian_cloud.cpp` Test 4 depends on it.
  Stride-0 with a non-zero vertex count remains an error.

Net delta: -555 +59 LOC. The new `ply_parse_core` itself is ~290 LOC
(header + cpp combined), so total deletion of duplicated logic is
~210 LOC — drift becomes structurally impossible for the parse path.

Validation:
- All 3 macOS configs build clean (debug / release / release-with-diag)
- `test_gaussian_cloud`, `test_ply2heightmap`, `test_vfx_lights_rotation`,
  `test_ply_path_resolution`, `test_bridge_reconnect_root`, and
  `test_gs_chunk_streamer` all pass on this branch
- Pre-existing failures (`test_gs_emission`, `test_scene_loading`,
  `test_incremental_sync`) confirmed identical on `origin/main`
- 8s `gseurat_demo` smoke (release): full pipeline loads
  (1,698,939 terrain Gaussians via runtime GSVX, 742,573 persistent
  dynamics, async load to 17 slabs), `[VFX] preload` produces the same
  `2000/50000 Gaussians (stride=25)` torch decimation as before, 0
  Vulkan validation errors

Follow-up (deferred per scope decision): the CI round-trip fixture
diff originally outlined as option (b) in issue #393. The shared lib
alone makes parse-core drift structurally impossible; the thin wrappers'
per-field conversions are still hand-written but the surface area is
trivial (one struct copy + one packing step).
